My_App/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
azure-storage-blob==12.18.3
numpy==1.26.4
pandas==2.2.2
xgboost==2.0.3
scikit-learn==1.4.2
joblib==1.3.2
